Progress being done on REST controller, bug in POST exists

+++ src/main/java/com/example/smartsofttesttask/Repositories/ProductRepository.java
@@ -4,5 +4,5 @@ import com.example.smartsofttesttask.Domains.Product;
 import org.springframework.data.repository.CrudRepository;
 
 public interface ProductRepository
-        extends CrudRepository<Product, Integer> {
+        extends CrudRepository<Product, Long> {
 }
