+++ src/main/java/com/example/smartsofttesttask/Repositories/PurchaseRepository.java
@@ -4,5 +4,5 @@ import com.example.smartsofttesttask.Domains.Purchase;
 import org.springframework.data.repository.CrudRepository;
 
 public interface PurchaseRepository
-        extends CrudRepository<Purchase, Integer>  {
+        extends CrudRepository<Purchase, Long>  {
 }
